Moved logging function into helper_fx file
