+++ fosquant/downloaddata.py
@@ -3,11 +3,12 @@ import getopt
 import os
 import subprocess
 import json
-import logging
 
 from datetime import datetime
 
 import pandas as pd
+
+from helper_fx import *
 
 sys.path.append("~/Github/fosquant/")
 
@@ -66,51 +67,7 @@ def check_existing_files(path_to_check, overwrite):
     else:
         return True
 
-def setup_logger(projectdir):
-    """Sets up logging by creating a logger object and making a directory if needed.
 
-    Use by calling logger.info() or logger.debug()
-
-    Args:
-        projectdir (Str): Path to folder where log directory will be created.
-
-    Returns:
-        logger: Object allowing lines to be added to log. 
-    """
-    logdir = os.path.join(projectdir, "log")
-    
-    if not os.path.isdir(logdir):
-        os.mkdir(logdir)
-
-    ## setting up logger
-    logfile = os.path.join(logdir, "{}.log".format(datetime.now().strftime('%Y-%m-%d_%H:%M:%S')))
-
-    logger = logging.getLogger(logfile)
-    logger.setLevel(level=logging.DEBUG)
-
-    logStreamFormatter = logging.Formatter(
-    fmt=f"%(levelname)-8s %(asctime)s \t line %(lineno)s - %(message)s", 
-    datefmt="%H:%M:%S"
-    )
-    consoleHandler = logging.StreamHandler(stream=sys.stdout)
-    consoleHandler.setFormatter(logStreamFormatter)
-    consoleHandler.setLevel(level=logging.DEBUG)
-
-    logger.addHandler(consoleHandler)
-
-    logFileFormatter = logging.Formatter(
-        fmt=f"%(levelname)s %(asctime)s L%(lineno)s - %(message)s",
-        datefmt="%Y-%m-%d %H:%M:%S",
-    )
-    fileHandler = logging.FileHandler(filename=logfile)
-    fileHandler.setFormatter(logFileFormatter)
-    fileHandler.setLevel(level=logging.DEBUG)
-
-    logger.addHandler(fileHandler)
-
-    logger.info("Created log file at {}".format(logfile))
-
-    return logger
 
 f = open("../config.json")
 config_data = json.load(f)
